Update nick.c - Add clarity to snotices
Added potential nick and ident to failed NICK snotice from trying to use a reserved nick. For cases where this happens pre-connection, it will say `<none>` in the snotice.

Additionally, the connection snotice for qlines was not clear and gave the impression that the name was being banned, and not an impression that the name had just connected.