Spamfilters with utf8 features in it where not always working and could result in a crash if you ran the 'SPAMFILTER' command as IRCOp.
Apparently not widespread since nobody reported it and I found it
while debugging another issue. Not sure when it was introduced,
I'm pretty sure it was working before. Or maybe there is/was some
second factor involved.

Anyway, fixed now and previous commit also adds extra code for in
case i screw up again with this, so it is not silently ignored.