Central Blocklist: fix issue with clients being killed if too slow. We already allow users in after 10 seconds if CBL is too slow, and that part worked correctly. However 5 seconds later, when the URL API 15 second timeout hits, it would try to allow the user in AGAIN. This caused the user to be introduced twice, causing remote servers to kill the user, and also screwing up user counts.
Reported by multiple people, including Jellis who actually filed
a report with logs, and alice providing logs as well, all when CBL
was temporarily down for a few hours in August 2024. In hindsight
this bug was already reported by k4be back in November 2023 but
was more rare at the time and mistaken for another very similar
bug that was fixed in 6.1.3.

With this patch, we check before we call cbl_allow(), but also
cbl_allow() itself checks the "user already allowed in?".

Oh yeah and this is hot patchable, within the hour I will make
this work: ./unrealircd hot-patch cbl-timeout-fix