Fix incorrect message about non-trusted SSL/TLS certificate when you use the default certificate/key (conf/tls/server.cert.pem) even when that cert is valid and issued by a trusted CA (like Let's Encrypt). You would get such an incorrect "best practices advice" on-boot, but (fortunately) not on each subsequent REHASH.
This was because the TLS system was not yet initialized completely at
the time of the best practices checks, ctx_server was NULL. This is
now solved by re-ordering some function calls.
This does change some win_error() and config_load_failed() stuff for
Windows so I hope that's okay.

Reported by Bun-Bun.