Include 'away' information in JSON-RPC users object. Reported/requested by CrazyCat: https://forums.unrealircd.org/viewtopic.php?p=40990 Inspired by Valware's PR: https://github.com/unrealircd/unrealircd/pull/319
This adds "away_reason" and "away_since". Note that the latter may not be as
reliable for remote users at the moment, because in case there was a split and
the server (re)connects, the away_since will be the time of the server resync
and not the original time that the user went away.