Make IsSynched() check if both the "far" server and the "near" server are synched. Both need to be checked, because: * The "far" server may be fully synched to "near" (and thus tagged as synced)   but the "near" server may be introducing the "far" server, when   we are connecting to "near" * The "near" server may be fully synched but the "far" server is connecting   in and may thus not be synched yet
In reality, things are even more complex, since one would have to verify
the whole chain of links. But.. yeah.

Long-story short: this fixes things like "User xyz joined #xxxxx" logging
where this showed up while the server was linking in. It is not supposed to
log that, similar to how we not log all 1000 users as newly connecting when
a 1000-user-server links in. In fact, it didn't already log that for
directly-connected-servers, but for far servers it did previously.

And... that again gave performance issues if you were connecting like a
100k-user far server.. since you suddenly had 100k * numchannels join events
being logged (which surprisingly still only took 6 seconds for 100k entries,
but still, it is wrong to do so and can be avoided).